requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
faker
transformers
openpyxl
sqlalchemy 
pymysql
pyodbc
